pyproject.toml
[project]
name = "spec-validator"
version = "0.1.0"
description = "Specification consistency audit and rule validation tooling."
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
